Major clean up of waitCmdListAddrAvailable
There bits of code in this routine, such as the NULL pointer check,
that were so obviously wrong that I can't believe they ever existed.
I have fixed it as best I can, however the test in the "No running
batch" case is just my best guess based on the existing comments.