VGA: Detect analog monitors differently from digital monitors
If a display is connected only to the DVI port, it can be
"accidentally" detected as both DVI and analog VGA.  If EDID is
supported, determine whether the EDID block says the monitor is
digital.  If EDID is not supported, do some voodoo to determine if
there is a monitor load present.